requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
tensorflow
python-multipart
requests
Pillow
aiofiles
